--- a/src/LightGraphs.jl
+++ b/src/LightGraphs.jl
@@ -151,6 +151,7 @@ and tutorials are available at the
 [JuliaGraphsTutorials repository](https://github.com/JuliaGraphs/JuliaGraphsTutorials).
 """
 LightGraphs
+include("utils.jl")
 include("interface.jl")
 include("deprecations.jl")
 include("core.jl")
@@ -207,7 +208,6 @@ include("flow/push_relabel.jl")
 include("flow/multiroute_flow.jl")
 include("flow/kishimoto.jl")
 include("flow/ext_multiroute_flow.jl")
-include("utils.jl")
 include("spanningtrees/kruskal.jl")
 include("spanningtrees/prim.jl")
 include("biconnectivity/articulation.jl")