--- a/src/community/cliques.jl
+++ b/src/community/cliques.jl
@@ -27,6 +27,7 @@ function maximal_cliques end
     T = eltype(g)
     # Cache nbrs and find first pivot (highest degree)
     maxconn = -1
+    # nnbrs = [Set{T}() for n in vertices(g)]
     nnbrs = Vector{Set{T}}()
     for n in vertices(g)
         push!(nnbrs, Set{T}())
@@ -55,7 +56,7 @@ function maximal_cliques end
     done = Set{T}()
     stack = Vector{Tuple{Set{T},Set{T},Set{T}}}()
     clique_so_far = Vector{T}()
-    cliques = Vector{Array{T}}()
+    cliques = Vector{Vector{T}}()
 
     # Start main loop
     while !isempty(smallcand) || !isempty(stack)
@@ -85,7 +86,7 @@ function maximal_cliques end
         end
         # Shortcut--only one node left!
         if isempty(new_done) && length(new_cand) == 1
-            push!(cliques, cat(1, clique_so_far, collect(new_cand)))
+            push!(cliques, vcat(clique_so_far, collect(new_cand)))
             pop!(clique_so_far)
             continue
         end