--- a/src/graphtypes/simplegraphs/simpledigraph.jl
+++ b/src/graphtypes/simplegraphs/simpledigraph.jl
@@ -117,11 +117,12 @@ end
 
 
 function rem_edge!(g::SimpleDiGraph, e::SimpleDiGraphEdge)
-    has_edge(g, e) || return false
-    i = searchsortedfirst(g.fadjlist[src(e)], dst(e))
-    deleteat!(g.fadjlist[src(e)], i)
-    i = searchsortedfirst(g.badjlist[dst(e)], src(e))
-    deleteat!(g.badjlist[dst(e)], i)
+    i = searchsorted(g.fadjlist[src(e)], dst(e))
+    isempty(i) && return false # edge doesn't exist
+    j = first(i)
+    deleteat!(g.fadjlist[src(e)], j)
+    j = searchsortedfirst(g.badjlist[dst(e)], src(e))
+    deleteat!(g.badjlist[dst(e)], j)
     g.ne -= 1
     return true
 end