--- a/src/graphtypes/simplegraphs/simpleedgeiter.jl
+++ b/src/graphtypes/simplegraphs/simpleedgeiter.jl
@@ -5,7 +5,7 @@ The function [`edges`](@ref) returns a `SimpleEdgeIter` for `AbstractSimpleGraph
 The iterates are in lexicographical order, smallest first. The iterator is valid for
 one pass over the edges, and is invalidated by changes to the graph.
 """
-mutable struct SimpleEdgeIter{T<:Integer} <: AbstractEdgeIter
+struct SimpleEdgeIter{T<:Integer} <: AbstractEdgeIter
     m::Int
     adj::Vector{Vector{T}}
     directed::Bool