--- a/src/operators.jl
+++ b/src/operators.jl
@@ -512,10 +512,10 @@ function merge_vertices!(g::Graph, vs::Vector{T} where T <: Integer)
 
     for i in vertices(g)
         # Adjust connections to merged vertices
-        if (i != merged_vertex) && !insorted(vs, i)
+        if (i != merged_vertex) && !insorted(i, vs)
             nbrs_to_rewire = Set{eltype(g)}()
             for j in out_neighbors(g, i)
-               if insorted(vs, j)
+               if insorted(j, vs)
                   push!(nbrs_to_rewire, merged_vertex)
                else
                  push!(nbrs_to_rewire, new_vertex_ids[j])
@@ -527,7 +527,7 @@ function merge_vertices!(g::Graph, vs::Vector{T} where T <: Integer)
         # Collect connections to new merged vertex
         else
             nbrs_to_merge = Set{eltype(g)}()
-            for element in filter(x -> !(insorted(vs, x)) && (x != merged_vertex), g.fadjlist[i])
+            for element in filter(x -> !(insorted(x, vs)) && (x != merged_vertex), g.fadjlist[i])
                 push!(nbrs_to_merge, new_vertex_ids[element])
             end