Some simplification of membership testing, and some VERY modest performance improvements.

(`foo()` just removes all edges in a graph.)

Master:
```
julia> @benchmark difference(h, g)
BenchmarkTools.Trial:
  memory estimate:  40.97 MiB
  allocs estimate:  464758
  --------------
  minimum time:     100.119 ms (14.32% GC)
  median time:      126.416 ms (13.82% GC)
  mean time:        153.632 ms (25.77% GC)
  maximum time:     295.614 ms (62.45% GC)
  --------------
  samples:          34
  evals/sample:     1


julia> @benchmark foo(g)
BenchmarkTools.Trial:
  memory estimate:  4.50 MiB
  allocs estimate:  26495
  --------------
  minimum time:     21.885 ms (0.00% GC)
  median time:      27.445 ms (0.00% GC)
  mean time:        30.885 ms (3.55% GC)
  maximum time:     99.003 ms (0.00% GC)
  --------------
  samples:          162
  evals/sample:     1
```
This PR
```
julia> @benchmark difference(h, g)
BenchmarkTools.Trial:
  memory estimate:  40.97 MiB
  allocs estimate:  464758
  --------------
  minimum time:     98.077 ms (5.44% GC)
  median time:      110.199 ms (14.50% GC)
  mean time:        124.564 ms (23.34% GC)
  maximum time:     249.726 ms (58.69% GC)
  --------------
  samples:          41
  evals/sample:     1


julia> @benchmark foo(g)
BenchmarkTools.Trial:
  memory estimate:  4.50 MiB
  allocs estimate:  26495
  --------------
  minimum time:     18.158 ms (0.00% GC)
  median time:      24.558 ms (0.00% GC)
  mean time:        27.837 ms (3.87% GC)
  maximum time:     107.658 ms (0.00% GC)
  --------------
  samples:          181
  evals/sample:     1
```