@sbromberger 2017-09-10
I plan on merging this sometime next week unless there are objections cc @jpfairbanks @juliohm @somil55 

also cc @mschauer