@mschauer 2017-09-10
In https://github.com/JuliaGraphs/LightGraphs.jl/blob/70a42f4a94a3929ed72082e50a6cdf355d10ea38/src/graphtypes/simplegraphs/simplegraph.jl#L148 try `isempty(i)` and `first(i)`?