@sbromberger 2017-09-10
> try `isempty(i)` and `first(i)`?

Yes. Thanks.